--- a/src/lineshape.cc
+++ b/src/lineshape.cc
@@ -2242,6 +2242,15 @@ struct Derivatives {
 //! Helper to keep function signature clean
 typedef Array<Derivatives> ArrayOfDerivatives;
 
+//! Helper function to find the last relevant derivative
+Index active_nelem(const ArrayOfDerivatives& derivs) noexcept {
+  return std::distance(derivs.cbegin(),
+                      std::find_if(derivs.cbegin(), derivs.cend(),
+                                    [](auto& deriv) {
+                                      return deriv.deriv == nullptr;
+                                    }));
+}
+
 struct ComputeValues {
   Complex * const F;
   Complex * const dF;
@@ -2249,7 +2258,9 @@ struct ComputeValues {
   Complex * const dN;
   const Numeric * const f;
   const Index size;
+  const ArrayOfDerivatives& derivs;
   const Index jac_size;
+  const Index max_jac_size;
   const bool do_nlte;
   
   Index jac_pos(Index iv, Index ij) const noexcept {return jac_size * iv + ij;}
@@ -2261,21 +2272,23 @@ struct ComputeValues {
                 const Vector &f_grid,
                 const Index start,
                 const Index nv,
-                const Index nj,
+                const ArrayOfDerivatives& derivs_,
                 const bool do_nlte_) noexcept :
   F(F_.get_c_array()+start),
-  dF(nj ? dF_.get_c_array()+start*nj : nullptr),
-  N(do_nlte_ ? N_.get_c_array()+start : nullptr),
-  dN((nj and  do_nlte_) ? dN_.get_c_array()+start*nj : nullptr),
+  dF(dF_.get_c_array()+start*derivs_.size()),
+  N(N_.get_c_array()+start),
+  dN(dN_.get_c_array()+start*derivs_.size()),
   f(f_grid.get_c_array()+start),
-  size(nv), jac_size(nj), do_nlte(do_nlte_) {}
+  size(nv), derivs(derivs_), jac_size(derivs_.size()), max_jac_size(active_nelem(derivs)), do_nlte(do_nlte_) {
+  }
   
   ComputeValues(Complex &F_, std::vector<Complex> &dF_,
                 Complex &N_, std::vector<Complex> &dN_,
                 const Numeric &f_lim,
-                const Index nj, const bool do_nlte_) noexcept :
-  F(&F_), dF(dF_.data()), N(&N_), dN(dN_.data()),
-  f(&f_lim), size(1), jac_size(nj), do_nlte(do_nlte_) {}
+                const ArrayOfDerivatives& derivs_, const bool do_nlte_) noexcept :
+  F(&F_), dF(dF_.data()), N(&N_), dN(dN_.data()), f(&f_lim), size(1),
+  derivs(derivs_), jac_size(derivs.nelem()), max_jac_size(active_nelem(derivs)), do_nlte(do_nlte_) {
+  }
   
   ComputeValues& operator-=(const ComputeValues& cut) ARTS_NOEXCEPT {
     ARTS_ASSERT(cut.size == 1, "Not a cutoff limit")
@@ -2283,15 +2296,15 @@ struct ComputeValues {
     
     for (Index iv=0; iv<size; iv++) {
       F[iv] -= *cut.F;
-      for (Index ij=0; ij<jac_size; ij++) {
-        dF[jac_pos(iv, ij)] -= cut.dF[ij];
+      for (Index ij=0; ij<max_jac_size; ij++) {
+        dF[jac_pos(iv, derivs[ij].jac_pos)] -= cut.dF[derivs[ij].jac_pos];
       }
     }
     if (do_nlte) {
       for (Index iv=0; iv<size; iv++) {
         N[iv] -= *cut.N;
-        for (Index ij=0; ij<jac_size; ij++) {
-          dN[jac_pos(iv, ij)] -= cut.dN[ij];
+        for (Index ij=0; ij<max_jac_size; ij++) {
+          dN[jac_pos(iv, derivs[ij].jac_pos)] -= cut.dN[derivs[ij].jac_pos];
         }
       }
     }
@@ -2586,16 +2599,16 @@ void cutoff_loop_sparse_linear(ComputeData &com,
   if ((dense_size + sparse_low_size + sparse_upp_size) == 0) return;
   
   // Get the compute data view
-  ComputeValues comval(com.F, com.dF, com.N, com.dN, com.f_grid, dense_start, dense_size, nj, do_nlte);
+  ComputeValues comval(com.F, com.dF, com.N, com.dN, com.f_grid, dense_start, dense_size, derivs, do_nlte);
   
   // Define the cutoff
   Complex Fc=0, Nc=0;
   std::vector<Complex> dFc(do_cutoff ? nj : 0, 0), dNc((do_cutoff and do_nlte) ? nj : 0, 0);
-  ComputeValues cut(Fc, dFc, Nc, dNc, fu, nj, do_nlte);
+  ComputeValues cut(Fc, dFc, Nc, dNc, fu, derivs, do_nlte);
   
   // Get views of the sparse data
-  ComputeValues sparse_low_range(sparse_com.F, sparse_com.dF, sparse_com.N, sparse_com.dN, sparse_com.f_grid, sparse_low_start, sparse_low_size, nj, do_nlte);
-  ComputeValues sparse_upp_range(sparse_com.F, sparse_com.dF, sparse_com.N, sparse_com.dN, sparse_com.f_grid, sparse_upp_start, sparse_upp_size, nj, do_nlte);
+  ComputeValues sparse_low_range(sparse_com.F, sparse_com.dF, sparse_com.N, sparse_com.dN, sparse_com.f_grid, sparse_low_start, sparse_low_size, derivs, do_nlte);
+  ComputeValues sparse_upp_range(sparse_com.F, sparse_com.dF, sparse_com.N, sparse_com.dN, sparse_com.f_grid, sparse_upp_start, sparse_upp_size, derivs, do_nlte);
   
   const Index nz = do_zeeman ? band.ZeemanCount(i, zeeman_polarization) : 1;
   for (Index iz = 0; iz < nz; iz++) {
@@ -2671,16 +2684,16 @@ void cutoff_loop_sparse_triple(ComputeData &com,
   if ((dense_size + sparse_low_size + sparse_upp_size) == 0) return;
   
   // Get the compute data view
-  ComputeValues comval(com.F, com.dF, com.N, com.dN, com.f_grid, dense_start, dense_size, nj, do_nlte);
+  ComputeValues comval(com.F, com.dF, com.N, com.dN, com.f_grid, dense_start, dense_size, derivs, do_nlte);
   
   // Define the cutoff
   Complex Fc=0, Nc=0;
   std::vector<Complex> dFc(do_cutoff ? nj : 0, 0), dNc((do_cutoff and do_nlte) ? nj : 0, 0);
-  ComputeValues cut(Fc, dFc, Nc, dNc, fu, nj, do_nlte);
+  ComputeValues cut(Fc, dFc, Nc, dNc, fu, derivs, do_nlte);
   
   // Get views of the sparse data
-  ComputeValues sparse_low_range(sparse_com.F, sparse_com.dF, sparse_com.N, sparse_com.dN, sparse_com.f_grid, sparse_low_start, sparse_low_size, nj, do_nlte);
-  ComputeValues sparse_upp_range(sparse_com.F, sparse_com.dF, sparse_com.N, sparse_com.dN, sparse_com.f_grid, sparse_upp_start, sparse_upp_size, nj, do_nlte);
+  ComputeValues sparse_low_range(sparse_com.F, sparse_com.dF, sparse_com.N, sparse_com.dN, sparse_com.f_grid, sparse_low_start, sparse_low_size, derivs, do_nlte);
+  ComputeValues sparse_upp_range(sparse_com.F, sparse_com.dF, sparse_com.N, sparse_com.dN, sparse_com.f_grid, sparse_upp_start, sparse_upp_size, derivs, do_nlte);
   
   const Index nz = do_zeeman ? band.ZeemanCount(i, zeeman_polarization) : 1;
   for (Index iz = 0; iz < nz; iz++) {
@@ -2786,13 +2799,13 @@ void cutoff_loop(ComputeData &com,
   if (not cutsize) return;
   
   // Get the compute data view
-  ComputeValues comval(com.F, com.dF, com.N, com.dN, com.f_grid, cutstart, cutsize, nj, do_nlte);
+  ComputeValues comval(com.F, com.dF, com.N, com.dN, com.f_grid, cutstart, cutsize, derivs, do_nlte);
   
   // Get the cutoff data view
   Complex Fc=0, Nc=0;
   std::vector<Complex> dFc(do_cutoff ? nj : 0, 0), dNc((do_cutoff and do_nlte) ? nj : 0, 0);
-  ComputeValues cut(Fc, dFc, Nc, dNc, fu, nj, do_nlte);
-
+  ComputeValues cut(Fc, dFc, Nc, dNc, fu, derivs, do_nlte);
+  
   const Index nz = do_zeeman ? band.ZeemanCount(i, zeeman_polarization) : 1;
   for (Index iz = 0; iz < nz; iz++) {
     const Numeric dfdH = do_zeeman ? band.ZeemanSplitting(i, zeeman_polarization, iz) : 0;
@@ -2913,7 +2926,7 @@ void line_loop(ComputeData &com,
         }
       }
     }
-    std::remove_if(begin(derivs), end(derivs), [](Derivatives& dd) { return dd.deriv == nullptr; });
+    std::remove_if(derivs.begin(), derivs.end(), [](Derivatives& dd) { return dd.deriv == nullptr; });
 
     // Call cut off loop with or without sparsity
     switch (speedup_type) {
@@ -2978,7 +2991,7 @@ void compute(ComputeData &com,
              const Options::LblSpeedup speedup_type) ARTS_NOEXCEPT {
   [[maybe_unused]] const Index nj = jacobian_quantities.nelem();
   const Index nl = band.NumLines();
-  [[maybe_unused]] const Index nv = com.f_grid.nelem();
+  const Index nv = com.f_grid.nelem();
 
   // Tests that must be true while calling this function
   ARTS_ASSERT(H >= 0, "Only for positive H.  You provided: ", H)