Here's my output from the script Oliver provided in a currently open issue:

Time per pressure/temperature combination
CASE 1 - All species w/ jacobian ArrayOfSpeciesTagVMR: 1.586376s
CASE 2 - All species w/ jacobian Line::VMR           : 3.782718s
CASE 3 - All species w/o jacobian                    : 1.323047s
CASE 4 - One species at a time                       : 1.330896s