@olemke 2021-04-28
I reran case 1 and 4 through the profiler. The remaining difference now comes down to the extra time spent in `ComputeValues::operator-=` for the derivates.

Case 1
<img width="1920" alt="case1_5000_4" src="https://user-images.githubusercontent.com/665724/116400859-d163bf00-a82a-11eb-9c09-b82929d74d16.png">

Case 4
<img width="1920" alt="case4_5000_4" src="https://user-images.githubusercontent.com/665724/116401002-053ee480-a82b-11eb-939c-3eae9a06d5fd.png">