@riclarsson 2021-04-28
Thank you for running the profiler, Oliver.  I will have a look.  It is definitely doing more than it should for case 1.  I think a similar cut-down in time should be possible as for frequency_loop, but I need to check.  (My initial quick-and-dirty stuff just now didn't solve it, so I need to have a more detailed look when there's time.)