@riclarsson 2021-04-29
New times are slightly faster.  For the same case as above (5000 frequencies):
```
Time per pressure/temperature combination
CASE 1 - All species w/ jacobian ArrayOfSpeciesTagVMR: 1.486118s
CASE 2 - All species w/ jacobian Line::VMR           : 3.820726s
CASE 3 - All species w/o jacobian                    : 1.339322s
CASE 4 - One species at a time                       : 1.349771s
```

Also for 100,000 frequencies:
```
Time per pressure/temperature combination
CASE 1 - All species w/ jacobian ArrayOfSpeciesTagVMR: 21.045467s
CASE 2 - All species w/ jacobian Line::VMR           : 64.153041s
CASE 3 - All species w/o jacobian                    : 20.579016s
CASE 4 - One species at a time                       : 20.542538s
```

In #339 the case 1 to 4 ratio was 150-200%.  It is 2-10% now.  The case 2 to 4 ratio is pretty much the same as before since it is computing the full derivative.