@riclarsson 2021-04-29
I'm marking this as ready for review since the previous tests showed where time was being spent.  I think the small 'regression' in case 2/3/4 is just noise.  They should not be seeing any difference that could cost time now cf the first time comparison of this PR (done at the commit before the latest)