@olemke 2021-04-29
I agree that the slight differences in case 2-4 are negligible. I see basically the same times for 3/4 and 2 scores even slightly lower.

```
Before cc84172
CASE 1: 1.273507
CASE 2: 2.751986
CASE 3: 1.070707
CASE 4: 1.074574

After cc84172
CASE 1: 1.129954
CASE 2: 2.709163
CASE 3: 1.071539
CASE 4: 1.075069
```